Avoid SSR randomization in tetromino previews
